Add Bencoding Int tests

+++ Netorrent/Bencoding/Structs/BInt.cs
@@ -1,10 +1,10 @@
 ﻿namespace Netorrent.Bencoding.Structs;
 
-public struct BInt(int data) : IBencodingType
+public struct BInt(long data) : IBencodingType
 {
-    public readonly int Data => data;
+    public readonly long Data => data;
 
-    public static implicit operator int(BInt other) => other.Data;
+    public static implicit operator long(BInt other) => other.Data;
 
-    public static implicit operator BInt(int data) => new(data);
+    public static implicit operator BInt(long data) => new(data);
 }
